updated misc
