+++ addons/acemisc/$PBOPREFIX$
@@ -1,1 +1,1 @@
-kat_acemisc
+x\kat_aceAdvMedical\addons\acemisc
