+++ addons/acemisc/CfgFunctions.hpp
@@ -1,0 +1,11 @@
+class CfgFunctions {
+    class ADDON {
+        tag = "ace_interaction";
+        class ace_interaction {
+            class addPassengersActions {
+                file = QPATHTOF(functions\fnc_addPassengersActions.sqf);
+            };
+        };
+    };
+
+};
