+++ addons/acemisc/CfgMoves.hpp
@@ -1,0 +1,17 @@
+class CfgMovesBasic {
+    class ManActions {
+        kat_stretcher = "kat_stretcher";
+    };
+};
+
+class CfgMovesMaleSdr: CfgMovesBasic {
+    class States {
+        class Crew;
+        class kat_stretcher: crew {
+            file = "a3\anims_f\data\anim\sdr\inj\ainjppnemstpsnonwnondnon.rtm";
+            interpolateTo[] = {"Unconscious", 0.02};
+            leftHandIKCurve[] = {};
+            rightHandIKCurve[] = {};
+        };
+    };
+};
