+++ addons/acemisc/XEH_init.sqf
@@ -1,0 +1,5 @@
+#include "script_component.hpp"
+
+params ["_unit"];
+
+_unit call FUNC(init);
