+++ addons/acemisc/XEH_preInit.sqf
@@ -1,0 +1,21 @@
+#include "script_component.hpp"
+
+ADDON = false;
+
+PREP_RECOMPILE_START;
+#include "XEH_PREP.hpp"
+PREP_RECOMPILE_END;
+
+#define CBA_SETTINGS_CAT "KAT - ADV Medical: Misc"
+
+//Enable aceExposure things
+[
+    QGVAR(enable),
+    "CHECKBOX",
+    localize LSTRING(SETTING_ENABLE),
+    CBA_SETTINGS_CAT,
+    [true],
+    true
+] call CBA_Settings_fnc_init;
+
+ADDON = true;
