+++ addons/acemisc/functions/fnc_attachStretcher.sqf
@@ -1,3 +1,4 @@
+#include "script_component.hpp"
 /*
  * Author: Katalam
  * Children action for attaching stretcher
@@ -9,7 +10,7 @@
  * None
  *
  * Example:
- * [cursorTarget] call kat_aceMisc_fnc_attachStretcher;
+ * [cursorTarget] call kat_acemisc_fnc_attachStretcher;
  *
  * Public: No
  */
@@ -23,7 +24,7 @@ private _actions = [];
 {
     private _type = typeOf _x;
     private _name = getText (configFile >> "CfgVehicles" >> _type >> "displayName");
-    private _uniqueName = format ["KAT_%1", _type];
+    private _uniqueName = format ["kat_acemisc_%1", _type];
     if (isArray (configFile >> "CfgVehicles" >> _type >> "stretcherPos")) then {
         _actions pushBack [
             [
