+++ addons/acemisc/functions/fnc_conditionIV.sqf
@@ -1,3 +1,4 @@
+#include "script_component.hpp"
 /*
  * Author: Katalam
  * Condition check for the medicine IV stands to remove a bag.
@@ -10,7 +11,7 @@
  * 0: Allowed <BOOLEAN>
  *
  * Example:
- * [player, 1000] call kat_aceMisc_fnc_conditionIV;
+ * [player, 1000] call kat_acemisc_fnc_conditionIV;
  *
  * Public: No
  */
@@ -20,11 +21,11 @@ params [
     ["_ml", 1000, [0]]
 ];
 
-if !(kat_aceMisc_enable) exitWith {false};
+if !(GVAR(enable)) exitWith {false};
 
 private _object = nearestObjects [_player, ['Land_IntravenStand_01_1bag_F', 'Land_IntravenStand_01_2bags_F'], 10, false];
 if (isNil "_object") exitWith {false};
-_object = _object select {_ml in (_x getVariable ["kat_aceMisc_stand", []])};
+_object = _object select {_ml in (_x getVariable [QGVAR(stand), []])};
 if (isNil "_object") exitWith {false};
 _object = _object select 0;
 if (_object distance _player <= 10) exitWith {true};
