+++ addons/acemisc/functions/fnc_removeIVbag.sqf
@@ -1,3 +1,4 @@
+#include "script_component.hpp"
 /*
  * Author: Katalam
  * Removes a IV bag for the medicine IV stands.
@@ -9,7 +10,7 @@
  * None
  *
  * Example:
- * [player, 500] call kat_aceMisc_fnc_removeIVbag;
+ * [player, 500] call kat_acemisc_fnc_removeIVbag;
  *
  * Public: No
  */
@@ -18,9 +19,9 @@ params ["_player", "_oldML"];
 
 // get some paramaters
 private _object = nearestObjects [_player, ['Land_IntravenStand_01_1bag_F', 'Land_IntravenStand_01_2bags_F'], 10, false];
-_object = _object select {_oldML in (_x getVariable ["kat_aceMisc_stand", []])};
+_object = _object select {_oldML in (_x getVariable [QGVAR(stand), []])};
 _object = _object select 0;
-private _value = _object getVariable ["kat_aceMisc_stand", []];
+private _value = _object getVariable [QGVAR(stand), []];
 private _className = typeOf _object;
 private _newObjectClass = "";
 if (_className isEqualTo "Land_IntravenStand_01_2bags_F") then {
@@ -36,7 +37,7 @@ if (_className isEqualTo "Land_IntravenStand_01_2bags_F") then {
 
 // creates new object at [0,0,0]
 private _newObject = createVehicle [_newObjectClass, [0,0,0], [], 0, "CAN_COLLIDE"];
-_newObject setVariable ["kat_aceMisc_stand", _value, true];
+_newObject setVariable [QGVAR(stand), _value, true];
 
 // replace the old with the new object
 private _oldPos = getPos _object;
