+++ addons/acemisc/models/stretcher/Model.cfg
