+++ addons/acemisc/script_component.hpp
@@ -1,0 +1,17 @@
+#define COMPONENT acemisc
+#define COMPONENT_BEAUTIFIED KAT - MISC
+#include "\x\kat_aceAdvMedical\addons\main\script_mod.hpp"
+
+// #define DEBUG_MODE_FULL
+// #define DISABLE_COMPILE_CACHE
+// #define ENABLE_PERFORMANCE_COUNTERS
+
+#ifdef DEBUG_ENABLED_MISC
+    #define DEBUG_MODE_FULL
+#endif
+
+#ifdef DEBUG_SETTINGS_MISC
+    #define DEBUG_SETTINGS DEBUG_SETTINGS_MISC
+#endif
+
+#include "\x\kat_aceAdvMedical\addons\main\script_macros.hpp"
